Refactor method visitors
